Things appear to be working
Using the code submitted by MyPaint to the gtk bugzilla I found a proper way to
avoid compression. Their code works. Now I need to update my code to
reflect their. I seem to be drawing in the wrong place. The proper way seems to be
first to collect the points, and then at draw callback ot draw the lines.
NOT to draw in the button press callback (as I do now).